changing file directories
so if you ever try to run this from main it is GOING to brick - with all the filetype reading that I saw no reason to refactor entirely I had to, instead, change the directories to look one more up, so there are a lot more "../" than there were before. If you were to want to run this from main for some forsaken reason, delete the "../" things you find in this commit's changes and go from there.