agent now works on both colors
so that works nicely - the one thing that I didn't like was that there wasn't an option to update the board within the original code as we went. It might be worth trying to push it so everyone doesn't have to rewrite it. Or maybe thats part of the point. I don't think that that is the most interesting part of the project, but maybe I am just boring. I would like to update my heuristic to be better, but I like the lookup table for right now.